v1.8.2
Add enemy bullet display

+++ projectile.py
@@ -1,21 +1,23 @@
 import pygame
 
+import alien
+
 from config import SCREEN
 from explosion import Explosion
 
 
 class Projectile:
-    def __init__(self, player):
+    def __init__(self, owner, x, y, vel):
         self.image = pygame.image.load("ammo.png")
         self.image = pygame.transform.scale(self.image, (self.image.get_width() // 4, self.image.get_height() // 4))
 
         self.rect = self.image.get_rect()
 
-        self.rect.x = player.x + 30
-        self.rect.y = SCREEN.get("height") - self.image.get_height() - 20
-        self.player = player
+        self.rect.center = (x + 30, y)
 
-        self.velocity = 8
+        self.owner = owner
+
+        self.velocity = vel
         # self.velocity = 40 # FOR TESTING
 
     def draw(self, surface):
@@ -23,24 +25,33 @@ class Projectile:
 
     def move(self):
 
-        for alien in self.player.aliens:
-            if alien.rect.colliderect(self.rect):
+        # If the bullet has been shot by an alien
+        if isinstance(self.owner, alien.Alien):
+            # If the bullet is out of the screen
+            if self.rect.y >= SCREEN.get('height'):
                 self.delete()
-                self.explode()
-                alien.destroy()
 
-        if self.rect.y <= 0 - self.image.get_height():
-            self.delete()
+        # If player shot a bullet
+        else:
+            for _alien in self.owner.aliens:
+                if _alien.rect.colliderect(self.rect):
+                    self.delete()
+                    self.explode()
+                    _alien.destroy()
 
+            if self.rect.y <= 0 - self.image.get_height():
+                self.delete()
+
+        # Decrement the bullet y coordinate by the defined amount of pixels in velocity variable.
         self.rect.y -= self.velocity
 
     def delete(self):
-        self.player.projectiles.remove(self)
+        self.owner.projectiles.remove(self)
 
     def explode(self):
         # TEST SPRITES
         # Create a Sprite object centered on x,y position
         explosion = Explosion(self.rect.x, self.rect.y)
         # Add the Sprite to the Group
-        self.player.explosions_grp.add(explosion)
-        self.player.sound.play('explosion')
+        self.owner.explosions_grp.add(explosion)
+        self.owner.sound.play('explosion')
